Handle multi-word arithmetic
These instructions are exceedingly rare, so we just make TCG fall back to the
slower implementations in terms of single-word arithmetic. The symbolic handling
would more or less follow that path anyway, and the low number of "mulu2" and
"muls2" that I see in real code doesn't really justify the effort of
implementing dedicated handling. Concrete execution of those instructions is
still very fast.