Use the garbage collector of the symbolic backend
At the end of a translation block, we don't have any local expressions left; the
only places where reachable expressions can be found are symbolic
memory (already known to and managed by the backend) and the region in the CPU
state where we keep the expressions for register values (communicated to the
backend on initialization). This is a good moment to collect expressions.