Update symcc submodule command in README
When symcc is a git submodule, running docker build without first
loading the submodules of symcc leads to an error.

The Dockerfile of symcc has a command that inits its submodules, but
this command does not work when the symcc directory is itself a
submodule (because in this case git tries to access the supermodule of
symcc, but it does not exist inside the docker image).